Changes to be committed: 	renamed:    lib/dimension/Dimension.js -> lib/dimension/dimension.js 	renamed:    lib/measure/Measure.js -> lib/measure/measure.js 	renamed:    lib/prefix/Prefix.js -> lib/prefix/prefix.js 	renamed:    lib/unit/Unit.js -> lib/unit/unit.js 	renamed:    test/measure/Measure.js -> test/measure/measure.js
