+++ lib/dimension/dimension.js
